# Make CPU Operator Stats Optional