Summary:
Similar to what https://github.com/facebookincubator/velox/pull/1534 did for Expressions, this adds a configuration property
"driver.track_operator_cpu_usage" to control whether cpu usage tracking for operators is on of off. The default is off.

We have seen in use cases with very small batches (128 or less) measuring the CPU time can itself account for ~10% of the CPU spent in processing
batches.

For example, running the SimpleArithmetic benchmark with batches of 100 rows

With CPU usage tracking:
```
multiplySmall                                               4.58ns   218.29M
multiplySameColumnSmall                                     4.37ns   228.67M
multiplyHalfNullSmall                                       7.86ns   127.17M
multiplyConstantSmall                                       5.08ns   196.67M
multiplyNestedSmall                                         6.80ns   147.16M
multiplyNestedDeepSmall                                    12.06ns    82.89M
```
Without CPU usage tracking:
```
multiplySmall                                               4.22ns   236.99M
multiplySameColumnSmall                                     3.99ns   250.70M
multiplyHalfNullSmall                                       7.12ns   140.39M
multiplyConstantSmall                                       4.67ns   214.27M
multiplyNestedSmall                                         5.94ns   168.39M
multiplyNestedDeepSmall                                    11.03ns    90.65M
```

Differential Revision: D39753640